BoxBacground rewritten to TS
