+++ src/components/BoxBackground/BoxBackground.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import BoxBackground from '@components/BoxBackground'
+import { BoxBackground } from '@components/BoxBackground'
 import Spacer from '@components/Spacer'
 
 # BoxBackground
